requirements.txt
accelerate==0.14.0
# clip==1.0
cmake==3.24.1.1
denoising-diffusion-pytorch==0.31.0
# dlib==19.24.0
einops==0.5.0
ema-pytorch==0.0.10
filelock==3.8.0
ftfy==6.1.1
gdown==4.5.3
gputil==1.4.0
grad-cam==1.4.5
joblib==1.2.0
lmdb==1.3.0
lpips==0.1.4
numpy==1.22.4
opencv-python==4.6.0.66
pytorch-fid==0.2.1
pyyaml==6.0
regex==2022.9.13
scikit-learn==1.1.2
threadpoolctl==3.1.0
torchattacks==3.3.0
torchviz==0.0.2
tqdm==4.64.1
ttach==0.0.3
cmake>=3.22.3
lmdb>=1.2.1
numpy>=1.19.2
Pillow>=8.4.0
PyYAML>=6.0
tqdm>=4.55.1
opencv_python>=4.5.2.52
ftfy>=6.0.3
regex>=2021.10.23
dlib>=19.22.1